Changed global vars to library state
The old PREPIC program used global variables for program-wide state.  This
state is now kept in the new ESCR_T that is allocated by the app.  In the
ESCR routines, this passed in as the first argument and called "e".

In this change, all the reference to global variables were edited to
reference inside the local e structure.