Created real subsystem
The main program is now just a thin shell that gets the command line
arguments and calls ESCR routines to do most of the remaining work of
executing the script code.

All PIC specific references were removed.  Hooks will have to be added
later so that applications can add these kinds of customizations
themselves.

Builds, but not run tested.