Exported limited app interface
The ESCR top level module is now a true client of the ESCR system.
Previously it included the private ESCR include file.  It now only
includes the public file.  Some routines were moved to the public file to
allow application access.

Added ability for the application to specify file name suffixes of main
files and include files.  Which suffix was used is now kept in the
per-file structure of each input file.