Made output file optional
The system is now in either "script" or "preprocessor" mode.  The
default is script mode.

In script mode, there is no default output file.  A output file can be
created by the application or the script code, but it is only written to
explicitly.  All input lines must be valid script command, comments, or
blank.

In preprocessor mode, it is assumed the script commands surround data
that is to be copied to the output file.  The purpose of the script is
to modify the input data before it is copied to the output file.  Input
lines that are not valid script command are assumed to be data lines and
are copied to the output file, although function in these lines are
expanded.