Added intrinsic function mechanism
The mechanism has been created so that functions are defined at initialization
and by user code.  Previously, a fixed list of functions was hard coded.

This version is work in progress that only defines the function "+" at
initialization, and that function always returns "42".  The general mechanism
can handle intrinsic function (defined by statically linked code) and
user-defined (defined with ESCR code at run time).  User-defined function
definition and execution has not been implemented yet.

The next step is to implement the previously existing functions as individual
routines, then add symbols for them during initialization.  The "+" function
is already added as a single example.

Routines for "small" functions are intended to go in the ESCR_IFUN module.  A
bunch of helper routines will need to be created to support easy implementation
of these functions.  Such helper routines were previously local in the single
routine that contained all the hard-coded functions.  The new helper routines
will be a little less convenient since they need to be external.  Each intrinsic
function will be a separate routine.