Changed intrinsic function interface
The invocation string of intrinsic functions is now kept in FUNSTR in the ESCR
state block.  This simplifies the interface to helper routines in intrinsic
function code.

This is a snapshot of work in progress on implementing the various intrinsic
functions that were previously hard-coded.  This effort must be put aside while
ESCR_TERM_GET and related routines are fixed to return STAT instead of bombing
directly on error.  The intrinsic function helper routines can't be written
until this is cleaned up.