Added STAT return to routines that get terms.
The routines that get terms in various forms now return a STAT parameter
instead of bombing on error as before.  All the code calling these
routines has been updated, and everything builds without error again.  Not
run tested.

The next task is to create the funtion helper routines that will be useful
to implement the various functions that were previously branched to from a
CASE statement in a single routine.  The next step after that is to
actually create the function routines using the new helper routines.

Functions are currently broken.  The only implemented function is "+" and
it just returns "42" as a test.