Created intrinsic function helper routines
Create the first pass of the intrinsic function helper routines.  These
are in the IFN module.  Everything builds, but these routines have not
been run tested yet.  The next step is to implement the intinsic functions
for real.

The intrinsic function interface was changed.  The function expansion
string is no longer passed to the function, and is instead in the state
block.  This simplifies the interface to intrinsic functions, and more
importantly, to some of the helper routines.