Implemented most string-handling functions.
Implemented many of the string-handling functions.  This included making a
few minor changes to the function mechanism, and adding some helper
functions in the IFN module.

The new file oldfunc.pas is the old source code for all the PREPIC
functions.  That function mechanism is now obsolete, but it still be
useful to look at the code.  This module is not part of the build.

The next task is to continue adding functions.  All the function routines
have been created in the IFUN module, and just return.  The real
executable code for each needs to be filled in.