Finished all intrinsic functions
All the intrinsic functions stubbed in last time have been filled in and
tested.  Added a few intrinsic function helper routines, and a bunch of
error status codes with messages.  Enhanced some of the existing error
status to include the value the error is about.

Removed OLDFUNC.PAS.  This was the source code for the old intrinsic
functions that were all in one routine.  This code is still available as
part of PREPIC.  Some functions in PREPIC were not implemented here
because they are specific to processing PIC code.  Those will be added by
the PREPIC application layer above this library in the new version of
PREPIC.