Fixed to work with PREPIC
A few fixes and enhancements were made resulting from changing PREPIC to
use this script system instead of implementing its own.  PREPIC seems to
work with this version, except that private intrinsic commands and
functions have not been implemented yet in PREPIC.

Tracking input files has been expanded to the more general tracking of
source code file snippets.  When a snippet is the whole file, as before,
it is in the list of input files as before.  Snippets can also exist
without being chained to this list.  These are intended to be given names
indexed from the new SRC symbol table.  This symbol table is initialized,
but no code uses it yet.