Intrinsic function helper routines to public include file
The declaration of all the intrinsic function helper routines were moved
to the public include file (the intrinsic function helper routines are
called ESCR_IFN_xxx).  This allows applications to add their own intrinsic
functions, which is something they are intended to be able to to.

Added intrinsic function helper routines to return raw characters, as
opposed to quoted strings.