Changed LOOP command
The loop command is now more flexible.  What were previously options that
had to be specified in a particular sequence are now arbitrary options
that can be specified in any sequence.  Merged all counted loop types into
one.  LOOP N is no longer a special case, for example.

Change FOR keyword to WITH, to more accurately reflect what it does,
despite convention of existing computing history.

The loop value is now available in a constant instead of a variable.