LOOP command fixups
Fixed debug messages accidentally left in LOOP command code last time.
Fixed internal links in the HTML documenation for the LOOP command.

Added IFUN.ES, which is a script template for performing some operation
for each intrinsic function.  This has no dedicated purpose.  The
per-function logic is intended to be re-written for the specific task at
hand.  This script was used, for example, to generate the initial HTML
template for describing the functions.