Added qualified symbols syntax
Symbols references are now generalized as name:type:ver.  TYPE and VER and
their leading colons are optional, so all old symbol references are legal.
TYPE explicitly selects the type of the symbol, like variable or
subroutine.  VER can be either a absolute 1-N version number, or relative
to the current version.

The VER and QUAL options were added to the SYM function.