Implemented arbitrary current version of symbols.
The data for a symbol in the symbol table is now a new descriptor that
points to the first, current, and last versions of the symbol.  It
previously just pointed to the current version.  This change allows the
current version to be more easily tracked separately, and allows there to
be no current version.

Execution blocks referenced from symbols by name (like subroutines or
macros), now explicitly save the current version of the name symbol on
creation, and restore it on deletion.  In the block, the current version
is the next older than the one referencing the block.

This means, for example, that calling a subroutine from inside itself
actually calls the next older version, not itself recursively.  Recursion
is possible by calling the "+1" version of the subroutine name.

This new symbol version handling was previously described in the
documentation.  This version of the software catches up to the
documentation.