LOOP BY default is now always 1.
The default increment of a counted loop is now always 1.  Previously, it
was 1 or -1 depending on whether the TO value was above or below FROM.
The original intent was that the loop would automatically count by 1 in
the direction of FROM to TO.

However, this was inconvenient for the much more common case where the
loop is to be executed N times, where N could be zero.  In the case of N =
0, the usual intent is to execute the loop 0 times, not with the values 1
and 0.  The new BY default allows it to work that way without having to
explicitly include "BY 1".

For example, the following will be executed 0 times (the loop code will
not be run at all):

  loop with ii from 1 to 0

Previously, the loop code would have been executed twice, with II set to 1
and 0.