Added RUN command, exit status option to STOP command
Added the RUN command.  This runs a external program, with the exit status
of that program written to the integer variable EXITSTATUS.

Added the exit status option to the STOP command.  This becomes the exit
status of the script.  The ESCR program now makes that its own exit status
code.