Added user-define functions.
Added commands FUNCTION and ENDFUNC to define functions, and the command
FUNCVAL to set the returned value of such functions.

The current parsing state can now be saved, new state used temporarily,
then the old state restored.  Each new state is tied to a execution block,
and is automatically removed with the previous state restored when that
execution block is deleted.

This changed touched much code.  The single parsing state was previously
variables directly in the library use context, ESCR_T.  Parsing state is
now wrapped into a new structure, PARSE_T.  A fixed PARSE_T is included in
ESCR_T, but it is accessed by PARSE_P in ESCR_T so that the current
parsing state can be easily switched.