Added LOOP DIR, work in progress.
Directory loops were added, but the DENT function to get the values of the
current loop iteration has not been implemented yet.  This version is work
in progress where information about the current directory entry is written
to standard output.

Basically, directory loops aren't really ready for use yet in this
version.