Added MAKE_ESCR_SCRIPTER.ESCR
This script sets up the Windows execution environment so that .escr
scripts can be run directly from the command line without explicitly
running the ESCR program or entering the file name suffix.  Put another
way, ESCR script can now be run the same as BAT scripts.

This script is intended to be run as part of the software installation
process.  The GET_UTIL_MIN script has been updated to include the ESCR
program, and to cause this script to be run when the software is
installed.