Fixed erroneous error after user-defined function return.
When user-define functions were invoked, a "extra token" error would
sometimes result, although there was no extra token.  This was due to the
function popping the parsing state, then the extra token check done in the
popped-to state, not the one the last command actually executed in.  This
check is now not done after a command that changes the parsing state.  All
such commands must check for extra tokens themselves, before releasing the
current execution block (which implicitly pops the parsing state).